Build Haskell side with Stack
Summary:

The Python side has an okay-ish build using an okay-ish shell script and
venvs. The Haskell side relied on ambient packages, which was not ideal.
Migrate it to using Stack (haskellstack.org) instead.

The Stack/Cabal config is trivial; the only thing worth mentioning is we
use the last LTS release of GHC 8 instead of the latest LTS release
because GHC 9 is currently broken in Stack on aarch64-osx
(https://gitlab.haskell.org/ghc/ghc/-/issues/20592).

Modify the Python side to use Stack if available (precisely: `stack` exists on
the `$PATH`). If present, it runs the Haskell side using `stack exec` (the user
is responsible for running `stack build` first.) Otherwise, use `runhaskell`
and rely on the user installing the required Haskell packages. If neither is
available, error.

Update the README.

Test plan:

With stack installed, run

    stack purge
    stack build
    ./util.sh run --input-file demos/website-excerpts.json --dump-solutions

Everything runs and is happy.  Modify the "is stack present" if statement in
solver.py to `if False`, `stack purge`, and run again; it still works.